requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
pymongo==4.6.0
openai==2.0.0
Flask-CORS==4.0.0
numpy==1.24.3

